LayoutTests/js/date-prototype-expected.txt

Check the Date's prototype properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Date.prototype.toGMTString is Date.prototype.toUTCString
PASS Date.prototype.toUTCString.name is "toUTCString"
PASS Date.prototype.toUTCString.length is 0
PASS typeof Date.prototype.toUTCString is "function"
PASS Object.getOwnPropertyDescriptor(Date.prototype, "toUTCString").configurable is true
PASS Object.getOwnPropertyDescriptor(Date.prototype, "toUTCString").enumerable is false
PASS Object.getOwnPropertyDescriptor(Date.prototype, "toUTCString").writable is true
PASS successfullyParsed is true

TEST COMPLETE
